app/pyproject.toml
[tool.poetry]
name = "Youttel-bot"
version = "0.1.0"
description = ""
authors = ["Erkin <erkinqara@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
pyTelegramBotAPI = "^3.7.7"
fastapi = "^0.79.0"
python-multipart = "^0.0.5"
pytube = "^12.1.0"
python-dotenv = "0.20.0"

[tool.poetry.dev-dependencies]
black = "^22.6.0"


[build-system]
requires = ["poetry-core>=1.1.0"]
build-backend = "poetry.core.masonry.api"
